Initial cut at automated Debian derived system for ARM
- Buildroot works, but it isn't really what I need or want for a debug
and development platform.
- Debian 12 has a lot more functionality, which is something I'm going
to want as my needs / desires want (e.g., I want to be able to compile
natively and that's really beyond the scope of what the buildroot folks
are trying to do).
- This requires end users to have a QEMU emulation setup that supports
transparent emulation. Note that there may also be problems when trying
to engage things like SSL from the `chroot` environment, so be aware on
what is happening.