Can now build a Debian derived root filesystem for the ARM board
- Plan is to create an SD card image that can be used to boot the board
- Next step is to automate creation of the /boot partition, put the
kernel, bootloaders, and device tree on the partition
- Need to suppress all the output in the second stage build script and
constrain it to the build script
- On device trees, I'm curious as to whether U-boot even needs my device
tree at compile time - the FSBL contains enough information to allow the
bootloader to boot.  Something to think about.

Ultimately what I want automated is enough to get a system up and
running with my own FSBL, kernel, U-Boot, and root filesystem, along
with a device tree of some sort that provides access to whatever is in
the PL that is expressed by the exported hardware design.