First check in where the build process works entirely for me
- Still don't have pushbutton SD images, which may be fine for a while
- Have to manually provide an FSBL ELF location still. This will be the
case until I write a Makefile that pulls the required source code from
the XSA and builds the ELF from Xilinx source code
- Will remain the user responsibility to manually configure and build
the kernel and U-boot which I believe is more than fair. If you are
unable to build those two from the source trees (and the tags) that are
indicated, then you're probably not the target audience here anyway
- Scripted device tree generation from the Xilinx source generator works
but it is entirely the user's responsibility to modify the .dtsi files
appropriately prior to compiling them into a flattened device tree blob.
- Bitstream generation is entirely manual still at this point for a
variety of reasons. Converting from a .bit to a .bin that is compatible
with the `zynq-fpga` driver in the kernel is still a manual process for
now.
- But the whole thing should work after that. Build your kernel, build
your bootloaders, build your hardware, make sure that the constants
are set correctly, and then push the button. The build directory will
have a bootfs.armhf and a rootfs.armhf directory that you can put on an
SD card and boot.  I'm probably not going to automate SD card
generation, at least not now.
- Boot parameters are still a bit of a mystery to me - the
/boot/extlinux/extlinux.conf file seems to take priority over anything
else, including what is in the device tree (good) but I can't seem to
change them on the command line (bad).  That's probably because the
extlinux file hasn't been loaded yet - maybe there's a way to halt that
process and intervene.  Otherwise, distro boot may not be the way that I
want to go and using a boot.scr that I can override from U-boot might be
a better way, at least for now.
- This is a good day! We've got what we wanted, learned a ton, and ready
to start hacking on some hardware I think.