First kernel module!
To build for ARM, need to

export ARCH=arm
export CROSS_COMPILE=arm-linux-gnueabihf-
export KDIR=<path-to-linux-xlnx-source>

Need to also have the cross compiler in the PATH (i.e., source the Vitis
settings64.sh script). And then, just run `make` from the directory
containing the Makefile for the module. The kernel kbuild system should
be engaged and generate the .ko module, which can then be copied to the
Microzed and loaded. Note that the kernel source in question needs to
actually be the one that was used to build the kernel. Also, if you
don't specify cross compilation, it will assume that your kernel hasn't
been prepared and it will launch into the basic command line config (I
cannot believe that is still in there).