Adding support to resize the final partition on first boot up.
When you first boot an SD card from this image, it will resize the root
partition to be whatever the remainder of the card supports. But the
EXT4 filesystem on that partition needs to be resized as well, so you'll
need to run something like `sudo resize2fs /dev/mmcblk0p2` in order to
actually take advantage of that extra space.