More pieces of my sbc device driver
- As I learn more, I'm starting to understand terms better - I think I'm
going to wind up with a single platform driver that presents the entire block RAM
devices to userspace and then probably another one that works with the
control register space, so it wouldnt surprise me if there are actually
more than one kernel modules at play here.
- I modified the device tree on the board so it only has one BRAM to
interact with because I was getting conflicts about more than one file
or something liek that.  My guess is that I'm going to have to handle
additional complexity of grouping or something like that in order for me
to get things like /dev/bram-0 and /dev/bram-1 to appear as device
nodes.
- Probably going to rename this driver so that it isn't as misleading as
'basic' anymore.