Module loads and creates entries in /dev for char devices
- Learned a lot during this process but concluding that making my own
character device driver from scratch is an unnnecessarily large hammer
- Using an existing kernel framework is going to be more effective, so
I'm going to switch my focus to using the uio framework instead
- I don't need DMA right now (maybe in the future, if it takes a lot of
time to load images, but I have my doubts that is a thing) and I don't
know if I'm ever going to need to generate interrupts to the degree that
interrupt latency is ever a concern.  I can see generating an interrupt
and halting the SBC if necessary so I can debug it, for sure, but that
isn't going to need anything resembling low latency.
- Still, this was a huge learnign experience - always plan to make two,
and throw the first (usually) in the trash.