A saner approach to managing the device tree sources.
- Going forward, I'm going to use this flow for developing the device
tree for the board:

XSA -> FSBL (from Vitis, for now, but later will use a Makefile and be
completely independent from the Xilinx toolstack). This is done by just
unzipping the exported archive file and compiling the FSBL source code
along with the configuration specific source code. Note that this will
completely put me in the driver seat when it comes to low level
configuration of the chip.  Goodbye Xilinx Tcl parsing bugs!!!

XSA -> device tree source include files (done from XSCT, using the
device tree source generator tool from Xilinx). There is a script in the
scripts directory that does this. Then copy the generated source code
here for whatever hacking is required.  Then compile it here via the
Makefile and the `system.dtb` can be sent to the board. No device tree
blobs or top level source files are required.