First cut at the EEPROM emulator is done.
- I haven't added the XCI file for the block RAM yet
- And it also occurrs to me that at some point I'm going to need to
start version controlling the project I've been working with on the side
as well as the IP that is used to make it
- I hate dealing with projects and revision control but I have spent a
tremendous amount of time trying to find ways to avoid using projects
and the IPI flow that Xilinx recommends and I'm not happy with anything
that I've come up with.
- I may just make a project, strip out all of the stuff that isn't
necessary, and then check that bare project in.