Added reset pin to AT28C256 emulator
- The chip itself doesn't have a reset and none is provided to the
fabric, but the block RAM controller may complain about not having a
reset during synthesis, so we add it here and the completely ignore it
in the RTL. Future application may need to use it.
- Added the top level module for the SBC core. This is the top level
that would be considered platform / architecture independent.
- In principle, this could be instantiated in any device without needing
a board or carrier around it. If you watned to put this in a Basys3, you
could, you would just need to put some sort of bridge between the block
RAM and the outside world (e.g., JTAG to AXI and then an AXI to BRAM
controller) and a clock source of some sort.