Fixed a port width mismatch that was driving me nuts.
- First, in order to get the correct bit width for the ports, you have
to specify the depth in the address map correctly.  The address editor
is what determines the width of the address bus in the block designer.
This is huge - when the address editor says "32K" or "8K" its not the
size of the block RAM.  It's the size in addresses!  So, if you have a
32K block RAM and want to address it in 4 byte chunks with 4 bit write
enables (quite the normal use case) then you have to change the size in
the address editor, or you'll get ports the wrong size.
- I suspect that since a lot of folks use Verilog to do all of this and
don't read their synthesis logs, they wind up with all manner of address
or data problems.
- Unclear if I'm going to have to modify the source code to the block
RAM C programs I've written or if it still supports byte widths and so
forth.