Addresses a memory bus width and depth issue for BRAM and device trees
- The map size that is in the /sys/class entry for block RAM describes
the size of the memory in accesses (e.g., a 32KB block is 8K of 32-bit
accesses)
- We wish the map size to reflect the size in bytes, because it is going
to get mapped, and seeked, and written, etc. as if it were an array of
bytes.  So, this needs to be fixed to reflect that.  Specifically, the
value read back from that needs to be shifted a bit so that its the
number of bytes rather than the number of accesses.
- This is the correct way to do it, but it may completely destroy all of
the existing code and the easiest way to know is to deply, build, and
try it.
- I anticipated this would happen when I started deploying the actual
block RAM that the board was going to use rather than just sticking
something in hardware and seeing if I could write it.  Now that I'm
past the proof of concept stage, I expected I would need to start
looking at this closer.
- If it breaks everything else entirely, then I suppose some rewrites
are going to be necessary.