Started working on adding a VIO / ILA combination to the ROM emulator
- It isn't clear to me the way addressing and byte access is done
at the two interfaces of the block RAM (at least, now that the actual
processor SBC block design is being implemented in hardware)
- The simplest way is to just add a VIO that can poke the memory in
hardware and dump it out to chipscope.
- I'm not trying to add a third port or a backdoor to the memory; that
already exists via the PS. I just need to verify that bytes go in and
out of memory the way that I think they do and that hasn't been shown to
my satisfaction yet.
- Once I've verified the hardware behavior, we can look at how to modify
the block RAM support software.